% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attributes.R
\name{add_attributes}
\alias{add_attributes}
\title{Append Data Attributes in Specified Column and Place}
\usage{
add_attributes(data, place, att, attcol, sep = ",")
}
\arguments{
\item{data}{A data frame containing a column of attributes to be updated}

\item{place}{Numeric. Which element to add `att` onto when `attcol` is
separated by `sep`.}

\item{att}{String. Attribute code to be appended.}

\item{attcol}{Name of column containing attributes.}

\item{sep}{String that separates attribute elements in `attcol`. Defaults to
`','`.}
}
\value{
Data frame with updated `attcol`.
}
\description{
Append Data Attributes in Specified Column and Place
}
\examples{

dat <- data.frame(
n = 1:3,
attributes = rep('all_,co2_,n2o_,ch4_,nh3_', 3)
)
add_attributes(dat, 1, '02', attributes)
}
