% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_rows.R
\name{add_missing_rows}
\alias{add_missing_rows}
\title{Add rows to data frame to make missing values explicit}
\usage{
add_missing_rows(
  data,
  site_var = site,
  plot_var = plot,
  date_var = Date,
  year_var = year,
  plots_unique_by = c(year, site),
  plots_unique_all = paste(treatment_key$year, treatment_key$site, treatment_key$plot),
  addl_id_vars = c(treatment, block)
)
}
\arguments{
\item{data}{A data frame}

\item{site_var}{Name of column in \code{data} containing site names. Defaults
to \code{site}.}

\item{plot_var}{Name of column in \code{data} containing plot IDs. Defaults
to \code{plot}.}

\item{date_var}{Name of column in \code{data} containing sampling dates.
Defaults to \code{Date}.}

\item{year_var}{Name of column in \code{data} containing sampling year.
Defaults to \code{year}.}

\item{plots_unique_by}{Names of columns in \code{data} within which plot IDs
are unique.  Defaults to \code{c(year, site)}.}

\item{plots_unique_all}{Vector of space-separated strings containing all
unique plots preceded by their corresponding levels of
\code{plots_unique_by}. Defaults to \code{paste(treatment_key$year,
treatment_key$site, treatment_key$plot)}.}

\item{addl_id_vars}{Additional names of ID columns (those not already listed
in other arguments) in \code{data}. The values of these variables will be
filled in added 'missing' rows in the result. Set to \code{NULL} if there
are no additional ID variables to be filled in. Defaults to c(treatment,
block).}
}
\value{
\code{data} with additional rows of previously missing levels of
  specified variable combinations.
}
\description{
Add rows to data frame to make missing values explicit
}
