% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions_plot_sub.R
\name{assumptions_plot_sub}
\alias{assumptions_plot_sub}
\title{Make Plot Subtitle Containing Levene and Shapiro-Wilk Test p-values}
\usage{
assumptions_plot_sub(
  data,
  var,
  trans_label,
  levene_label = "levene",
  shapiro_label = "shapiro",
  var_then_label = TRUE,
  sep = "_"
)
}
\arguments{
\item{data}{Data frame containing columns with Levene and Shapiro p values.
There must be only one unique p-value for each test within the data frame
or current group.}

\item{var}{Column name of variable for which ANOVA assumptions are being
tested. E.g. \code{co2}}

\item{trans_label}{Optional. String describing data transformation performed
on \code{var} prior to assumptions checks. E.g. 'log'.}

\item{levene_label}{String pattern which is appended on column names
containing the Levene test p-value. If \code{var_then_label} is
\code{TRUE} (the default), the name of the Levene p-value column should be
\code{paste0(var, sep, levene_label)}. Defaults to.}

\item{shapiro_label}{String pattern which is appended on column names
containing the Shapiro-Wilk test p-value. If \code{var_then_label} is
\code{TRUE} (the default), the name of the Shapiro p-value column should be
\code{paste0(var, sep, shapiro_label)}. Defaults to \code{"shapiro"}.}

\item{var_then_label}{Logical. Are column names containing p-values formatted
so that the name specified in \code{var} comes before the label specified
in \code{levene_label} and \code{shapiro_label}, separated by \code{sep}?
Defaults to \code{TRUE}.}

\item{sep}{Character pattern separating \code{var} from \code{levene_label}
or \code{shapiro_label} in names of columns containing p-values. Defaults
to "_".}
}
\value{
A string that can be used as a formatted plot subtitle.
}
\description{
Make Plot Subtitle Containing Levene and Shapiro-Wilk Test p-values
}
