% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_new_col.R, R/attr_show_codes.R,
%   R/attr_update.R
\name{attr_new_col}
\alias{attr_new_col}
\alias{attr_show_codes}
\alias{attr_update}
\title{Data Attributes}
\usage{
attr_new_col(
  data,
  attr_var = attributes,
  prefixes = c("all", "co2", "n2o", "ch4", "nh3"),
  sep = ","
)

attr_show_codes()

attr_update(data, attr_var = attributes, prefix, attr_code, sep = ",")
}
\arguments{
\item{data}{A data frame}

\item{attr_var}{Name of column in \code{data} where attributes are stored.
Defaults to \code{attributes}.}

\item{prefixes}{Character. Vector of prefixes to assign in new attribute
column with \code{attr_new_col()}. Defaults to \code{c('all', 'co2',
'n2o', 'ch4', 'nh3')}.}

\item{sep}{String that separates attribute elements in \code{attr_var}.
Defaults to \code{','}.}

\item{prefix}{Character. Unique prefix preceding attribute codes to be
updated with \code{attr_update()}.}

\item{attr_code}{Character. Attribute code to be appended with
\code{attr_update()}.}
}
\value{
\code{attr_new_col()} and \code{attr_update()} return a data frame
  with new or updated \code{attr_var}, respectively. \code{attr_show_codes()}
  returns a message to the console.
}
\description{
Data attributes are used in this workflow to record the
  reason(s) certain data points are marked for exclusion from modeling. These
  attributes are recorded for each gas in a single column, usually named
  \code{attributes}, created with \code{\link{attr_new_col()}}.
  \code{attr_update()} is used to add attributes to this column.
  \code{attr_show_codes()} prints each unique attribute code and its meaning
  to the console.
}
\examples{

attr_show_codes()
dat <- data.frame(x = rnorm(3))
dat <- attr_new_col(dat)
attr_update(dat, prefix = 'co2', attr_code = '02')
}
