% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gas_lmer.R
\name{build_gas_lmer}
\alias{build_gas_lmer}
\title{Build linear mixed effect models for GHG data}
\usage{
build_gas_lmer(data, formula, gas_vars = c(co2, n2o, ch4, nh3))
}
\arguments{
\item{data}{A data frame}

\item{gas_vars}{Names of columns in \code{data} containing gas data to be
substituted in \code{formula}. Each name should be an unquoted expression.
Defaults to \code{c(co2, n2o, ch4, nh3)}.}
}
\value{
A list of models
}
\description{
Build \code{\link[lmerTest]{lmer}()} models on multiple gas variables
simultaneously.
}
