% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_manual.R
\name{build_manual}
\alias{build_manual}
\title{Build and Prep Manual for Export}
\usage{
build_manual()
}
\value{
Nothing in memory.
}
\description{
For internal use only. Builds manual to inst/instructions and renames it to
match version annotation of pptm instructions.
}
