% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gas_loads.R
\name{calculate_gas_loads}
\alias{calculate_gas_loads}
\title{Calculate cumulative per-day gas loads from imputed data}
\usage{
calculate_gas_loads(
  data,
  gas_vars = c(co2_kg_ha_day_imputed, n2o_kg_ha_day_imputed, ch4_kg_ha_day_imputed,
    nh3_kg_ha_day_imputed),
  date_var = Date,
  site_var = site,
  plot_var = plot,
  method = "AUC",
  by_hour = FALSE
)
}
\arguments{
\item{data}{A data frame}

\item{gas_vars}{Names of columns in \code{data} containing imputed (i.e.
containing no NA values) gas data in mg kg\eqn{^{-1}}. Defaults to
\code{c(co2_kg_ha_day_imputed, n2o_kg_ha_day_imputed,
ch4_kg_ha_day_imputed, nh3_kg_ha_day_imputed)}.}

\item{date_var}{Name of column in \code{data} containing sampling dates in
YYYY-MM-DD format. Defaults to \code{Date}.}

\item{site_var}{Name of column in \code{data} containing site names. Defaults
to \code{site}.}

\item{plot_var}{Name of column in \code{data} containing plot IDs. Defaults
to \code{plot}.}

\item{method}{Defaults to \code{'AUC'} (area under the curve), currently the
only option. Linear interpolation methods to be added.}
}
\value{
\code{data} with new columns for gas loads
}
\description{
Calculate cumulative per-day gas loads after imputing gas data.
  This function calculates cumulative load using area under the curve
  ("AUC"). Each gas data point is assumed to be the hourly load for all days
  surrounding the measurement, up to one half of the distance between the
  measurement and its nearest preceding and following measurement.

  Additional methods, like linear interpolation, may be added in the future
  as options in the \code{method} argument.
}
