% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_seconds_lm_optim_helpers.R
\name{calib_optim}
\alias{calib_optim}
\title{Optimization of Carbon Dioxide ppm per Second Models}
\usage{
calib_optim(x, data, excl_var, min_n = 4)
}
\arguments{
\item{x}{Vector of \code{0}s and \code{1}s which will be randomly generated
in \code{\link{calib_range}()}}

\item{data}{Data frame containing CO\eqn{_2} data}

\item{excl_var}{Logical column in \code{data} which tracks observations that
should be excluded from the model}
}
\value{
Number of data points excluded from the model
}
\description{
Finds optimum values of \code{excl_var} column in \code{data} to
  make the R\eqn{^2} value of the \code{\link{lm}()} performed in
  \code{\link{ppm_seconds_lm_optim}()} at least 0.98 while minimizing the amount of data
  points excluded from the model.
}
\details{
See \code{\link{model_co2}()} for details about the parameters and
  method of model optimization.
}
\seealso{
\code{\link{model_co2}()}, \code{\link{calib_range}()}
}
