% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_seconds_lm_optim_helpers.R
\name{calib_range}
\alias{calib_range}
\title{Optimization of Carbon Dioxide ppm per Second Models}
\usage{
calib_range(data, excl_var, min_n = 4, x)
}
\arguments{
\item{data}{Data frame containing CO\eqn{_2} data}

\item{excl_var}{Logical column in \code{data} which tracks observations that
should be excluded from the model}

\item{min_n}{Integer. Minimum number of data points to retain in final model.
Defaults to 4.}

\item{x}{Passed from \code{\link{calib_optim}}; not used in this function.
Removing this parameter causes an error in \code{\link[stats]{optim}}, but
it does not need to be explicitly specified when the function is called.}
}
\value{
Vector of \code{0}s and \code{1}s with the same length as
  \code{data}.
}
\description{
Generates random instances of the logical column specified in
  \code{excl_var}, contained in \code{data}. In each random instance, existing
  \code{TRUE} observations are retained, and a random number of points at the
  head and tail of the column are replaced with \code{TRUE}. Maximum number
  of \code{TRUE} observations in the final instance is set at \code{min_n}.
  Converts randomized vector to integer for compatibility with
  \code{\link[stats]{optim}()}.
}
\seealso{
\code{\link{calib_optim}}, \code{\link{model_co2}}
}
