% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{exclude_outliers}
\alias{exclude_outliers}
\title{Mark outliers for exclusion}
\usage{
exclude_outliers(
  data,
  gas_var,
  excl_var,
  attr_var = attributes,
  prefix,
  sep = ",",
  n_IQR = 3,
  datatable = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{gas_var}{Name of column in \code{data} containing gas flux data to be
screened for outliers.}

\item{excl_var}{Name of logical column in \code{data} which is equal to
\code{TRUE} when an observation of \code{gas_var} should be excluded from
downstream processing and analysis.}

\item{attr_var}{Passed to \code{\link{attr_update}()}. Name of column in
\code{data} which records attributes. Defaults to \code{attributes}.}

\item{prefix}{Unique prefix denoting attribute element in \code{attr_var} to
be updated with \code{attr_update()}. If not specified, it will be inferred
from \code{gas_var}.}

\item{sep}{Passed to \code{\link{attr_update}()}. String that separates
attribute elements in \code{attr_var}. Defaults to \code{','}.}

\item{n_IQR}{Integer. Passed to \code{\link{is_outlier}}. Number of
interquartile ranges below 25% quantile or above 75% quantile at which a
value will be considered an outlier. Defaults to 3.}

\item{datatable}{Logical. Should a datatable (using
\code{\link[DT]{datatable}()} of extracted negative flux values be printed?
Defaults to \code{FALSE}.}
}
\value{
\code{data} with updated \code{excl_var} and \code{attr_var}.
}
\description{
Determine outliers using \code{\link{is_outlier}()} and mark
  them for exclusion from downstream processing.
}
