% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{Make formula with sprintf-style replacements}
\usage{
f(formula, repl, verbose = FALSE)
}
\arguments{
\item{formula}{String to be converted to formula, with instances of \code{\%s}
replaced by elements of \code{repl}}

\item{repl}{character vector of replacements for each instance of \code{\%s}
in \code{formula}}

\item{verbose}{Logical. Should replaced formula be printed to the console?
Defaults to \code{FALSE}.}
}
\value{
A formula with specified replacements
}
\description{
Shortcut for, e.g. \code{as.formula(sprintf(\%s ~ x, "y"))}. To
  be used inside a modeling function call.
}
\examples{
x <- rnorm(20)
y1 <- rnorm(20)
y2 <- rnorm(20)

# Single replacement
formula <- '\%s ~ x'
lm(f(formula, 'y1'))
lm(f(formula, 'y2'))

# Mutliple replacements
formula2 <- 'x ~ \%s * \%s'
lm(f(formula, c('y1', 'y2')))

}
