% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasmet_bind_check.R
\name{gasmet_bind_check}
\alias{gasmet_bind_check}
\title{Check that Gasmet data list is ready for binding}
\usage{
gasmet_bind_check(dfls, issues_path = "gasmet_data_import_issues.txt")
}
\arguments{
\item{dfls}{List of data frames to check before binding}

\item{issues_path}{File path to .txt file generated with
\code{\link{gasmet_file_issues}()}. Defaults to
'gasmet_data_import_issues.txt'}
}
\value{
TRUE if data frames in \code{dfls} are ready for binding
}
\description{
Performs several checks that evaluate if \code{dfls} is ready for binding
with \code{\link[dplyr]{bind_rows}}. Gives errors with specifics if binding
will be unsuccessful. Otherwise, the output of the function is not useful.
}
