% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_instructions.R
\name{get_latest_instructions}
\alias{get_latest_instructions}
\title{Get Latest Version of PowerPoint Package Instructions}
\usage{
get_latest_instructions(directory = "instructions")
}
\arguments{
\item{directory}{String. Directory in which to copy PowerPoint file. Defaults to \code{"instructions"}.}
}
\value{
Nothing in memory. Copies latest instructions to \code{directory} if the latest version is not already contained there. If the latest version is already present, a message is printed to the console.
}
\description{
Get Latest Version of PowerPoint Package Instructions
}
