% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_sum_stats.R
\name{ghg_sum_stats}
\alias{ghg_sum_stats}
\title{Summary Statistics for GHG Data}
\usage{
ghg_sum_stats(.data, ..., showDT = TRUE)
}
\arguments{
\item{.data}{A data frame containing GHG data}

\item{...}{Variable(s) in \code{.data} to be summarized}

\item{showDT}{Logical. Do you want to print an HTML data table of results
made with \code{\link[DT]{datatable}()}? Defaults to \code{TRUE}}
}
\value{
A data frame of summary statistics
}
\description{
Shortcut function to get calculate mean, median, min, max, standard
deviation, standard error, and count for \code{sum_vars}. Best used in
combination with \code{\link[dplyr]{group_by}()}. See examples.
}
\details{
Columns in results will have the format "{.col}_{.fn}" where
  "{.col}" is the original column name in \code{data} and "{.fn}" is the
  function that has been performed in the summary. See examples.
}
\examples{
data <- data.frame(
  co2 = rnorm(100, mean = 1000),
  n2o = rnorm(100, mean = 1000),
  treatment = rep(1:4, 25),
  site = rep(1:2, each = 50)
)
data \%>\% group_by(treatment, site) \%>\% ghg_sum_stats(co2, n2o, showDT = FALSE)

}
