% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_prefix.R
\name{guess_prefix}
\alias{guess_prefix}
\title{Infer prefix from a gas column name}
\usage{
guess_prefix(based_on)
}
\arguments{
\item{based_on}{Name of gas column}
}
\value{
Prefix compatible with attributes
}
\description{
Infer prefix from a gas column name
}
