% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute missing gas data}
\usage{
impute_missing_values(
  data,
  gas_vars = c(co2_kg_ha_day, n2o_kg_ha_day, ch4_kg_ha_day, nh3_kg_ha_day),
  excl_vars = NULL,
  site_var = site,
  date_var = Date,
  plot_var = plot,
  keep_all_cols = FALSE
)
}
\arguments{
\item{data}{A data frame}

\item{gas_vars}{Names of columns in \code{data} containing gas data to be
imputed. Defaults to \code{c(co2_kg_ha_day, n2o_kg_ha_day, ch4_kg_ha_day,
nh3_kg_ha_day)}.}

\item{excl_vars}{Names of logical columns in \code{data} which are
\code{TRUE} when the corresponding \code{gas_vars} should be replaced with
NA and then imputed. The order of names should match those in
\code{gas_vars} (i.e. the first column specified in \code{gas_vars} is
imputed when the first column specified in \code{excl_vars} is \code{TRUE}.
Defaults to \code{c(co2_exclude, n2o_exclude, ch4_exclude, nh3_exclude)}.}

\item{site_var}{Name of column in \code{data} containing site names. Defaults
to \code{site}.}

\item{date_var}{Name of column in \code{data} containing sampling dates.
Defaults to \code{Date}.}

\item{plot_var}{Name of column in \code{data} containing plot IDs. Defaults
to \code{plot}.}

\item{keep_all_cols}{Logical. Should intermediate columns used in imputation
calculations be retained in the output?}
}
\value{
\code{data} with new columns for imputed values. When
  \code{keep_all_cols} is \code{TRUE}, intermediate columns generated for
  imputation calculations (e.g. grand means by plot) are retained in the
  output.
}
\description{
Replaces \code{gas_vars} with \code{NA} according to
  \code{excl_vars}. Then fills NAs with imputed values using the procedure
  described in \code{\strong{Details}}.
}
\details{
Imputation is performed using the following procedure:

\enumerate{

 \item{
   Data are arranged by \code{site_var}, \code{plot_var}, and \code{date_var}
   so that each unique plot's samplings appear in chronological order in
   \code{data}.
 }

 \item{
   Columns are added that fill NAs with the nearest non-NA observation before
   AND after each missing value.
 }

 \item{
   The distance between the missing values and each of the values it was
   filled with is calculated (i.e. how many samplings between the missing
   value and the nearest non-NA values before and after, chronologically?)
 }

 \item{
   If the minimum distance between the missing value and a non-NA value is 2
   or less, the missing value is replaced with the mean of the previous
   non-NA value and the next non-NA value chronologically.
 }

 \item{
   If the minimum distance is greater than 2, the missing value is replaced
   with the grand mean of all non-NA observations for the plot.
 }

}
}
