% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_n.R
\name{include_n}
\alias{include_n}
\title{Count Observations Not Yet Excluded from Processing}
\usage{
include_n(data, exclude)
}
\arguments{
\item{data}{A data frame containing a logical column tracking processing
exclusions by observation.}

\item{exclude}{Name of logical column which tracks processing exclusions.
Defaults to `exclude_obs`.}
}
\value{
Integer count of rows in `data` where `exclude` is `FALSE`.
}
\description{
Count Observations Not Yet Excluded from Processing
}
\examples{

dat <- data.frame(
x = rnorm(10),
exclude_obs = sample(c(TRUE, FALSE), 10, replace = TRUE)
)
include_n(dat)

}
