% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{is_outlier}
\alias{is_outlier}
\title{Detect outliers}
\usage{
is_outlier(x, n_IQR = 3)
}
\arguments{
\item{n_IQR}{Integer. Number of interquartile ranges below 25% quantile or
above 75% quantile at which a value will be considered an outlier. Defaults
to 3.}

\item{gas_var}{Gas flux column in which to detect outliers}
}
\value{
A logical vector the same length as x (or current group)
}
\description{
Intended for use inside \code{\link[dplyr]{mutate}()} or
  \code{\link[dplyr]{filter}()}. Detects outliers in column or grouped column
  \code{x} as values less than the 25% quantile minus three interquartile
  ranges or greater than the 75% quantile plus three interquartile ranges.
  Returns \code{TRUE} for outliers and \code{FALSE} for non outliers.
}
