% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_preprocessing_report.R
\name{make_preprocessing_report}
\alias{make_preprocessing_report}
\title{Perform Preprocessing Steps and Make HTML Reports}
\usage{
make_preprocessing_report(
  data,
  date_var = Date,
  site_var = site,
  rmd_file = "code/110_preprocessing.Rmd",
  output_data_dir,
  output_report_dir,
  output_file,
  parent = rstudioapi::getActiveDocumentContext()$path,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame containing a single unique date and site of sampling}

\item{date_var}{Name of column in \code{data} containing the date of
sampling. Defaults to \code{Date}.}

\item{site_var}{Name of column in \code{data} containing the site of
sampling. Defaults to \code{site}.}

\item{rmd_file}{RMarkdown file from which to generate reports using
\code{\link[rmarkdown]{render}()}. Defaults to
\code{"code/110_preprocessing.Rmd"}.}

\item{output_data_dir}{Optional. Directory in which to store data outputs. If
not specified, data outputs will be stored in
\code{"data/01_converted_to_flux_rate/site_var_1"}, where \code{site_var_1}
is replaced with the name of the single unique site in data$site_var. The
directory will be created if it does not already exist.}

\item{output_report_dir}{Optional. Directory in which to store HTML report
outputs. If not specified, reports will be stored in
\code{"data/01_converted_to_flux_rate/site_var_1/html_reports"}, where
\code{site_var_1} is replaced with the name of the single unique site in
data$site_var. The directory will be created if it does not already exist.}

\item{output_file}{Optional. Name of file outputs. If not specified, files
will be named \code{"YYYYMMDD_preprocessing"}, where \code{"YYYYMMDD"} is
replaced with the single unique date in \code{data$date_var}.}

\item{overwrite}{Logical. Should \code{\link[rmarkdown]{render}()} be
performed if \code{output_file}.html already exists at
\code{output_report_dir}? Defaults to \code{FALSE}.}

\item{verbose}{Logical. Should messages be printed to the console? Defaults
to \code{TRUE}.}
}
\value{
Nothing in R memory. Exports reports and data files to specified
  directories.
}
\description{
Perform Preprocessing Steps and Make HTML Reports
}
