% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_modeling.R
\name{model_co2}
\alias{model_co2}
\title{Perform CO~2~ Modeling with Optimization}
\usage{
model_co2(data, exclude, co2_col = Carbon.dioxide.CO2)
}
\arguments{
\item{data}{Data frame containing CO~2~ data}

\item{exclude}{Logical column in `data` which tracks observations that should
be excluded from the model}
}
\value{
A tibble with the same number of rows as `data` and two columns: one
matching the name specified in `exclude`, containing optimized values, and
the other named `rsq` containing the R^2^ value of the optimized model.
}
\description{
Uses `calib_optim()` and `calib_range()` functions to model CO~2~ data with
optimization as described in `?calib_optim`.
}
