% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_levene_shapiro_p.R
\name{mutate_levene_shapiro_p}
\alias{mutate_levene_shapiro_p}
\title{Add levene and shapiro p values to model data}
\usage{
mutate_levene_shapiro_p(
  data,
  formula,
  gas_vars = c(co2, n2o, ch4, nh3),
  resid_vars = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{gas_vars}{Names of columns in \code{data} containing gas data to be
substituted in \code{formula}. Each name should be an unquoted expression.
Defaults to \code{c(co2, n2o, ch4, nh3)}.}

\item{resid_vars}{Unquoted expressions. Names of columns in \code{data}
containing model residuals. Order should correspond with \code{gas_vars}.
If NULL, \code{gas_vars} will be appended with "_resid" to deduce the
residual variable names.}
}
\value{
\code{data} with \code{length(gas_vars) * 2} new columns
}
\description{
Add levene and shapiro p values to model data
}
