% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_residuals.R
\name{mutate_residuals}
\alias{mutate_residuals}
\title{Add model residuals to data}
\usage{
mutate_residuals(data, models, suffix = "_resid")
}
\arguments{
\item{data}{A data frame}

\item{models}{List of models generated with \code{\link{build_gas_lmer}}.}

\item{suffix}{String. By default, names of new columns containing residuals will be based on the names of the models in \code{models} (usually gas variable names), plus a suffix of \code{"_resid"}. A different suffix can be specified with this argument.}
}
\value{
\code{data} with \code{length(models)} new columns.
}
\description{
Add model residuals to data frame used in modeling. Typically used after \code{\link{build_gas_lmer}}.
}
