% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_temp_fig.R
\name{noaa_temp_fig}
\alias{noaa_temp_fig}
\title{Daily Minimum and Maximum Air Temperatures Barplot}
\usage{
noaa_temp_fig(
  data,
  date_var = Date,
  tmax = NOAA_airtemp_c_max,
  tmin = NOAA_airtemp_c_min,
  site_var = NULL,
  sample_dates,
  bar_fill_colors = c("#D2B4DE", "#AED6F1", "#A569BD", "#3498DB"),
  date_breaks = "1 month",
  date_labels = "\%b \%Y",
  text_size = 14,
  x_lab = "Date",
  y_lab = "Daily Minimum and Maximum\\nAir Temperature (°C)"
)
}
\arguments{
\item{data}{A data frame containing one row per calendar day}

\item{date_var}{Name of column in \code{data} containing dates in
\code{"YYYY-MM-DD"} format. Defaults to \code{Date}.}

\item{tmax}{Name of column in \code{data} containing daily maximum air
temperature. Defaults to \code{NOAA_airtemp_c_max}.}

\item{tmin}{Name of column in \code{data} containing daily minimum air
temperature. Defaults to \code{NOAA_airtemp_c_min}.}

\item{site_var}{Name of column in \code{data} containing site names. If
provided, it will be used in the \code{rows} argument of
\code{\link[ggplot2]{facet_grid}()}. Defaults to \code{NULL}.}

\item{sample_dates}{Vector of all unique dates of sampling. By default, bars
on these dates will be darker than non-sampling-date bars.}

\item{bar_fill_colors}{Vector of length 4 specifying colors for,
respectively, daily max temp on a non-sampling date, daily min temp on a
non-sampling date, daily max temp on a sampling date, and daily min temp on
a sampling date. Defaults to \code{c('#D2B4DE', '#AED6F1', '#A569BD',
'#3498DB')}.}

\item{date_breaks}{String. Passed to \code{\link[ggplot2]{scale_x_date}()}.
Defaults to \code{'1 month'}.}

\item{text_size}{Integer. Size of text on plots. Defaults to \code{14}.}

\item{x_lab}{String or expression. X-axis label. Defaults to \code{"Date"}.}

\item{y_lab}{String or expression. Y-axis label. Defaults to \code{"Daily
Minimum and Maximum\nAir Temperature (°C)"}.}
}
\value{
A ggplot object
}
\description{
Daily Minimum and Maximum Air Temperatures Barplot
}
