% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_seconds_lm.R
\name{ppm_seconds_lm}
\alias{ppm_seconds_lm}
\alias{ppm_seconds_lm_optim}
\alias{ppm_seconds_lm_asis}
\title{Model Gas ppm per Second}
\usage{
ppm_seconds_lm(
  data,
  gas_var,
  excl_var = exclude_obs,
  min_n = 4,
  seconds_var = seconds,
  attr_var = attributes,
  prefix,
  optimization = FALSE,
  optim_method = 2
)

ppm_seconds_lm_optim(
  data,
  gas_var = Carbon.dioxide.CO2,
  excl_var = exclude_obs,
  min_n = 4,
  seconds_var = seconds,
  attr_var = attributes,
  prefix = "all",
  optimization = TRUE,
  optim_method = 2
)

ppm_seconds_lm_asis(
  data,
  gas_var,
  min_n = 4,
  excl_var = exclude_obs,
  seconds_var = seconds,
  attr_var = attributes,
  prefix,
  optimization = FALSE
)
}
\arguments{
\item{data}{Data frame containing CO\eqn{_2} data}

\item{gas_var}{Name of column in \code{data} containing gas ppm data to be
modeled. Defaults to \code{Carbon.dioxide.CO2} when \code{optimization =
TRUE}.}

\item{excl_var}{Logical column in \code{data} which is equal to \code{TRUE}
when an observation of \code{gas_var} should be excluded from modeling and
downstream analyses. Defaults to \code{exclude_obs}.}

\item{min_n}{Integer. Minimum number of data points for optimization
algorithm to retain in final model. Passed to \code{\link{calib_optim}()}
and \code{\link{calib_range}()}. Defaults to 4.}

\item{seconds_var}{Name of column in \code{data} containing seconds since
start of gas measurement. Defaults to \code{seconds}.}

\item{attr_var}{Name of column in \code{data} where attributes are stored.
Defaults to \code{attributes}.}

\item{prefix}{Passed to \code{\link{attr_update}()}. Defaults to \code{"all"}
when \code{optimization = TRUE}. If missing when \code{optimization =
FALSE}, it will attempt to be inferred from \code{gas_var}.}

\item{optimization}{Logical. Whether or not to perform optimization algorithm
descibed in \code{\strong{Details}}. Defaults to \code{FALSE}.}

\item{optim_method}{Double equal to \code{1} or \code{2}. Method \code{2} is
recommended and the default. Method \code{1} is the legacy method. Ignored
when \code{optimization = FALSE}.}
}
\value{
\code{data} with new columns containing model coefficients and
  updated \code{attr_var}.
}
\description{
Create a model of gas ppm per second, with or without use of an
  optimization algorithm to remove endpoints attributable to Gasmet warm-up
  or cool-down.

  When \code{optimization = TRUE}, the function uses an algorithm to model
  CO\eqn{_2} data with optimization as described in
  \code{\strong{'Details'}}. Returns \code{data} with optimized
  \code{excl_var} column, updated \code{attr_var}, and additional columns:
  \code{co2_rsq}, \code{co2_intercept}, and \code{co2_slope}, which
  correspond to the R\eqn{^2}, intercept, and slope of the final model,
  respectively.

  When \code{optimization = FALSE} (the default), the function creates a
  linear model of \code{gas_var ~ seconds_var} excluding observations where
  \code{excl_var == TRUE}. If the model R\eqn{^2} is at least 0.1, the
  function adds the model coefficients and R\eqn{^2} value in new columns in
  \code{data}. Otherwise, the same columns are added but with all entries
  \code{NA}. Attributes are updated in \code{data} as needed (see
  \code{?\link{attr_show_codes}} for details). A logical column is added to
  track exclusions of the specific gas being modeled.

  Typically, this function should be run first on CO\eqn{_2} data with
  \code{optimization = TRUE}, then on other gases with \code{optimization =
  FALSE}.
}
\details{
When \code{optimization = TRUE}, and \code{optim_method = 1} the
  optimization proceeds as follows:

\enumerate{

\item If \code{data} has less than \code{min_n} observations where
\code{excl_var == FALSE}, modeling is not performed and '02' is added to
\code{data$attr_var} for all gases. New columns are added with all values
\code{NA}.

\item Otherwise, \code{\link[stats]{optim}()} is called:

    \itemize{

       \item Random instances of \code{data$excl_var} are generated using
       \code{\link{calib_range}()}:

       \itemize{

           \item Existing observations where \code{data$excl_var == TRUE} are
           retained
           \item n observations at the head (n\eqn{\mathrm{_{head}}}) and/or
           tail (n\eqn{\mathrm{_{tail}}}) of data$excl_var are replaced with
           \code{TRUE}, where \code{n = 0 - m}, and \code{m} is a dynamic
           integer which keeps the sum of \code{n\eqn{\mathrm{_{head}}}} and
           \code{n\eqn{\mathrm{_{tail}}}} less than \code{nrow(data) - min_n}
           (i.e. the number of observations included remains
           \code{\eqn{\geq} \code{min_n}})

       }

       \item Each random instance of \code{data$excl_var} is fed to
       \code{\link{calib_optim}()} which executes the following scheme:

       \itemize{

         \item A model of CO\eqn{_2} per second, excluding observations
         marked for exclusion in the random instance, is generated
         \item If the model R\eqn{^2} value is \eqn{\geq} 0.98, the result
         fed to optim is the number of observations excluded from the model
         (this is the value we want to minimize)
         \item If the model R\eqn{^2} value is < 0.98, an arbitrary,
         extremely large value is returned to discourage
         \code{\link[stats]{optim}()} from choosing that model
         \item The previous three steps are repeated 1000 or
         \code{nrow(data) * 10} times, whichever is larger.

       }
       }

   \item If there is never a model with R\eqn{^2} \eqn{\geq} 0.98 in any
   iteration, the modeling is considered a failure case, all values in
   \code{data$excl_var} are updated to \code{TRUE}, '03' is added to
   \code{data$attributes}, and new columns are added with all values
   \code{NA}.

   \item Otherwise, the model is considered a success case and the resulting
   data frame contains the values of \code{data$excl_var} used in the final
   model, and new columns with model coefficients.
}

  When \code{optimization = TRUE}, and \code{optim_method = 2} the
  optimization utilizes successive linear models and exclusion of maximum
  magnitude residuals (at endpoints of the line only) until a model is built
  with R\eqn{^2} \eqn{\geq} 0.98 or until the number of observations
  remaining in the model is less than \code{min_n}.
}
\seealso{
\code{\link{ppm_seconds_plot}()}, \code{\link{calib_optim}()},
  \code{\link{calib_range}()}
}
