% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_seconds_plot.R
\name{ppm_seconds_plot}
\alias{ppm_seconds_plot}
\title{Create ggplot for ppm Gas per Second Linear Models}
\usage{
ppm_seconds_plot(
  data,
  gas_var,
  gas_name = c("co2", "n2o", "ch4", "nh3"),
  excl_var = paste0(gas_name, "_exclude"),
  rsq_var = paste0(gas_name, "_rsq"),
  co2_excl_var = exclude_obs,
  co2_rsq_var = co2_rsq,
  plot_var = plot,
  seconds_var = seconds
)
}
\arguments{
\item{data}{A data frame}

\item{gas_var}{Name of column in \code{data} containing gas data to be
plotted}

\item{gas_name}{Character. Name of gas that should be used in messages on
plots. Options are \code{'co2'}, \code{'n2o'}, \code{'ch4'}, and
\code{'nh3'}. Should correspond to \code{gas_var}.}

\item{excl_var}{Name of logical column in \code{data} which tracks
observations of \code{gas_var} which were excluded from ppm per second
modeling. Defaults to \code{paste0(gas_name, "_excl")}.}

\item{rsq_var}{Name of column containing R-squared values for the model of
\code{gas_var} ~ \code{seconds_col}. Defaults to \code{paste0(gas_name,
"_rsq")}.}

\item{co2_rsq_var}{Name of column containing R-squared value for CO2 ppm ~
seconds model. If \code{\link{model_co2}()} was used, this should be
\code{co2_rsq}. Defaults to \code{co2_rsq}.}

\item{plot_var}{Name of column containing plot IDs. Defaults to \code{plot}.}

\item{seconds_var}{Name of column containing number seconds since the start
of the gas measurement in the plot. Defaults to \code{seconds}.}
}
\value{
A plot
}
\description{
Create side-by-side ggplots of pre- and post-optimization models
  of gas ppm per second, or a plot displaying a failure message when models
  do not meet quality control parameters.
}
\details{
\code{data} should contain data for a single plot on a single
  sampling date.
}
\seealso{
\code{\link{model_co2}()}
}
