% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplots}
\alias{qqplots}
\title{Make Normal Quantile-Quantile Plots for Multiple Gases}
\usage{
qqplots(
  data,
  gas_vars = c(co2, n2o, ch4, nh3),
  resid_vars = NULL,
  one_plot = TRUE
)
}
\arguments{
\item{data}{A data frame}

\item{gas_vars}{Unquoted expressions. Names of columns in \code{data}
containing gas data.}

\item{resid_vars}{Unquoted expressions. Names of columns in \code{data}
containing model residuals. Order should correspond with \code{gas_vars}.
If NULL, \code{gas_vars} will be appended with "_resid" to deduce the
residual variable names.}

\item{one_plot}{Logical. Should plots be arranged on a single grid in the
output? If \code{FALSE}, a list of plots is returned.}
}
\value{
If \code{one_plot = TRUE}, a single plot. If \code{one_plot = FALSE},
  a list of ggplots.
}
\description{
Make Normal Quantile-Quantile Plots for Multiple Gases
}
