% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_fig.R
\name{regression_fig}
\alias{regression_fig}
\title{Scatterplot Figure with Regression Line, Equation, and Stats}
\usage{
regression_fig(
  data,
  x_var,
  y_var,
  group_var,
  facet_rows_var = gas,
  facet_cols_var = site,
  labeller = NULL,
  group_colors,
  x_lab = "",
  y_lab = "",
  color_lab = "",
  text_size = 14,
  point_shape = 16,
  point_alpha = 0.5,
  legend_position = "top",
  eqn_x = 0,
  eqn_y = Inf,
  eqn_vjust = 1.5,
  rr_x = 0,
  rr_y = Inf,
  rr_vjust = 1
)
}
\arguments{
\item{rr_vjust}{}
}
\value{
A plot
}
\description{
Scatterplot Figure with Regression Line, Equation, and Stats
}
