% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_gas_with_na.R
\name{replace_gas_with_na}
\alias{replace_gas_with_na}
\title{Replace gas variable entries with NA based on their corresponding exclusion
tracking columns}
\usage{
replace_gas_with_na(
  data,
  gas_vars = c(co2_kg_ha_day, n2o_kg_ha_day, ch4_kg_ha_day, nh3_kg_ha_day),
  excl_vars = c(co2_exclude, n2o_exclude, ch4_exclude, nh3_exclude)
)
}
\arguments{
\item{data}{A data frame}

\item{gas_vars}{Names of columns in \code{data} containing gas data to be
replaced with NA when the corresponding variable in \code{excl_vars} is
\code{TRUE}. Defaults to \code{c(co2_kg_ha_day, n2o_kg_ha_day,
ch4_kg_ha_day, nh3_kg_ha_day)}.}

\item{excl_vars}{Names of logical columns in \code{data} which are
\code{TRUE} when the corresponding \code{gas_vars} should be replaced with
NA. The order of names should match those in \code{gas_vars} (i.e. the
first column specified in \code{gas_vars} is replaced with NA when the
first column specified in \code{excl_vars} is \code{TRUE}. Defaults to
\code{c(co2_exclude, n2o_exclude, ch4_exclude, nh3_exclude)}.}
}
\value{
\code{data} with updated \code{gas_vars} columns
}
\description{
Replace gas variable entries with NA based on their corresponding exclusion
tracking columns
}
