% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_random_fx.R
\name{rm_random_fx}
\alias{rm_random_fx}
\title{Remove random effects from lmer-type model formula}
\usage{
rm_random_fx(formula)
}
\arguments{
\item{formula}{String. Formula of the type \code{"y ~ x + (1|block)"}.}
}
\value{
String containing only fixed effects from formula
}
\description{
Convenience function to remove random effects from \code{\link[lme4]{lmer}}
type model formulas, which is sometimes necessary for model checking
functions like Levene's test.
}
