% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_p.R
\name{round_p}
\alias{round_p}
\title{Round P Values for Plot Labels}
\usage{
round_p(p)
}
\arguments{
\item{p}{Numeric. P value to be rounded}
}
\value{
Character \code{"< 0.001"} or \code{"= p"} where \code{p} is rounded
  to 3 decimal places.
}
\description{
Rounds \code{p} to three decimal places. If the rounded value is less than
0.001, the output is \code{"< 0.001"}. Otherwise, the output is
\code{"= 0.000"} where \code{0.000} is the rounded \code{p}.
}
\examples{

round_p(0.000025038)
round_p(0.424610370)
}
