% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_boxplots_flux.R
\name{summary_boxplots_flux}
\alias{summary_boxplots_flux}
\title{Summary Boxplots of Daily Flux Data}
\usage{
summary_boxplots_flux(
  data,
  x_var = treatment,
  y_vars = c(co2_kg_ha_day, n2o_kg_ha_day, ch4_kg_ha_day, nh3_kg_ha_day),
  by = c(site, Date),
  x_lab = "Treatment",
  y_labs = c("Carbon Dioxide\\n(kg CO2-C/ha/day)", "Nitrous Oxide\\n(kg N2O-N/ha/day)",
    "Methane\\n(kg CH4-C/ha/day)", "Ammonia\\n(kg NH3-N/ha/day)"),
  box_color = "sienna2",
  box_fill = "transparent",
  box_alpha = 1,
  point_shape = 21,
  point_alpha = 0.5,
  point_color = "black",
  point_fill = "gray",
  plot_var = plot,
  facet_scales = "free_y"
)
}
\arguments{
\item{data}{A data frame}

\item{x_var}{Name of column in \code{data} that should be plotted as the x
variable. Defaults to \code{treatment}.}

\item{y_vars}{Name(s) of column(s) in \code{data} that should be plotted as
the y variable(s). If mutliple names are given, plots will be faceted so
that there is one facet per y variable. See
\code{\link[ggplot2]{facet_grid}()} for details.}

\item{by}{Split the data by these variables and make a plot for each. Defaults to \code{c(site, Date)}. Set to \code{NULL} to generate a single plot.}

\item{x_lab}{String. Label that will appear on the x axis. Defaults to
\code{'Treatment'}.}

\item{y_labs}{String. Label(s) that will appear on the y axis or facet
labels. Order should correspond to \code{y_vars}.}

\item{box_color}{String. Outline color of boxes in boxplots. Defaults to
\code{'sienna2'}.}

\item{box_fill}{String. Fill color of boxes in boxplots. Defaults to
\code{"transparent"}.}

\item{box_alpha}{Numeric 0-1. Opacity of boxes in boxplots. Defaults to
\code{1}.}

\item{point_shape}{Integer. Shape of points on boxplots. Defaults to
\code{21}.}

\item{point_alpha}{Numeric 0-1. Opacity of points on boxplots. Defaults to
\code{0.5}.}

\item{point_color}{String. Outline color of points on boxplots. Defaults to
\code{'black'}.}

\item{point_fill}{String. Fill color of points on boxplots. Defaults to
\code{"gray"}.}

\item{plot_var}{Name of column in \code{data} containing plot IDs. Defaults
to \code{plot}.}

\item{facet_scales}{String. Passed to \code{scales} arg of
\code{\link[ggplot2]{facet_grid}()}. Defaults to \code{"free_y"}.}
}
\value{
A single boxplot if \code{by} is NULL, or else a list of boxplots.
}
\description{
Summary Boxplots of Daily Flux Data
}
