% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series}
\alias{time_series}
\title{Time Series of Gas Data}
\usage{
time_series(
  data,
  date_var = Date,
  mean_var = gas_value,
  se_var = se,
  lines_group_var = treatment,
  facet_rows_var = gas,
  facet_cols_var = site,
  labeller = "label_value",
  line_colors,
  x_lab = "Date",
  y_lab = "",
  color_lab = "Treatment",
  date_breaks = "1 month",
  date_labels = "\%b \%Y",
  text_size = 14
)
}
\arguments{
\item{data}{A data frame. Should be summarized (in most cases, by
site/Date/treatment).}

\item{date_var}{Name of column in \code{data} containing sampling date in
\code{"YYYY-MM-DD"} format. Defaults to \code{Date}.}

\item{mean_var}{Name of column in \code{data} containing mean gas value to be
used as the y variable. Defaults to \code{gas_value}.}

\item{se_var}{Name of column in \code{data} containing standard error of the
mean gas value to be used in error bars. Defaults to \code{se}.}

\item{lines_group_var}{Name of column in \code{data} to use as grouping
variable for lines. Defaults to \code{treatment}. Set to \code{1} to plot a
single line.}

\item{facet_rows_var}{Name of column in \code{data} to be used in faceting
rows using \code{\link[ggplot2]{facet_grid}()}. Defaults to \code{gas}. Set
to \code{NULL} to remove row faceting.}

\item{facet_cols_var}{Name of column in \code{data} to be used in faceting
columns using \code{\link[ggplot2]{facet_grid}()}. Defaults to \code{site}.
Set to \code{NULL} to remove column faceting.}

\item{labeller}{Passed to \code{\link[ggplot2]{labeller}()}. A labeller
object containing specifications to be used in
\code{\link[ggplot2]{facet_grid}()}. See \code{?\link[ggplot2]{facet_grid}}
for details.}

\item{line_colors}{Vector of colors to be used for lines. Length must match
length of unique levels of \code{lines_group_var} in \code{data}. If not
specified, the function will use 'Set3' from
\code{\link[RColorBrewer]{brewer.pal}()}.}

\item{x_lab}{Label for the x axis. Defaults to \code{'Date'}.}

\item{y_lab}{Label for the y axis. Defaults to empty.}

\item{color_lab}{Label for color legend. Defaults to \code{"Treatment"}.}

\item{date_breaks}{Passed to \code{\link[ggplot2]{scale_color_manual}()}.
Defaults to \code{'1 month'}.}

\item{text_size}{Size of text on plot. Defaults to \code{14}.}
}
\value{
A plot
}
\description{
Time Series of Gas Data
}
