% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_modeling.R
\name{calib_optim}
\alias{calib_optim}
\title{Optimization of CO~2~ ppm per Second Models}
\usage{
calib_optim(x, data, exclude)
}
\arguments{
\item{x}{Vector of `0`s and `1`s which will be randomly generated in
`calib_range()`}

\item{data}{Data frame containing CO~2~ data}

\item{exclude}{Logical column in `data` which tracks observations that should
be excluded from the model}
}
\value{
R^2^ * -1; negative because `optim()` minimizes output
}
\description{
Finds optimum values of `exclude` column in `data` to maximize the R^2^ value
of the `lm()` performed in `model_co2()`.
}
