% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_modeling.R
\name{calib_range}
\alias{calib_range}
\title{Optimization of CO~2~ ppm per Second Models}
\usage{
calib_range(data, exclude, x)
}
\arguments{
\item{data}{Data frame containing CO~2~ data}

\item{exclude}{Logical column in `data` which tracks observations that should
be excluded from the model}

\item{x}{Passed from `calib_optim()`; not used in this function. Removing
this parameter causes an error in `optim()`, but it does not need to be
explicitly specified when the function is called.}
}
\value{
Vector of `0`s and `1`s with the same length as
`nrow(data)`.
}
\description{
Generates random instances of the logical column specified in `exclude`,
contained in `data`. In each random instance, existing `TRUE` observations
are retained, and a random number of points at the head and tail of the
column are replaced with `TRUE`. Maximum number of `TRUE` observations is set
at the number of existing `TRUE` values in the vector minus 4. Converts
randomized vector to integer for compatibility with `optim()`.
}
