% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasmet_file_issues.R
\name{gasmet_file_issues}
\alias{gasmet_file_issues}
\title{Detect Issues that Will Affect Gasmet TXT File Import}
\usage{
gasmet_file_issues(
  dfls,
  check_co2 = TRUE,
  check_n2o = TRUE,
  check_ch4 = TRUE,
  check_nh3 = TRUE,
  check_co = TRUE,
  ...
)
}
\arguments{
\item{dfls}{List of data frames, each list item containing data from a single
site/date}

\item{check_co2}{Logical. Whether or not to check issues in CO~2~ related
columns. Defaults to `TRUE`.}

\item{check_n2o}{Logical. Whether or not to check issues in N~2~O related
columns. Defaults to `TRUE`.}

\item{check_ch4}{Logical. Whether or not to check issues in CH~4~ related
columns. Defaults to `TRUE`.}

\item{check_nh3}{Logical. Whether or not to check issues in NH~3~ related
columns. Defaults to `TRUE`.}

\item{check_co}{Logical. Whether or not to check issues in CO related
columns. Defaults to `TRUE`.}

\item{...}{}
}
\value{
Message with details about fatal and non-fatal issues. Writes to the
console by default. When `file` is specified, writes to that file.
}
\description{
Detects issues in TXT files from the Gasmet such as missing columns,
mismatches between dates in data columns and file/folder names, multiple
dates contained in the same file, etc. Uses `cat()` to write a message
with details about detected issues (and if they need to be resolved for
processing to continue). Does not fix the issues - that must be done manually
by the user.
}
