% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chamber_volume.R
\name{import_chamber_volume}
\alias{import_chamber_volume}
\title{Import Chamber Volume Data}
\usage{
import_chamber_volume()
}
\value{
A data frame of all chamber volume data.
}
\description{
Imports and cleans up chamber volume Excel spreadsheets contained in
`./data/raw_data/chamber_volume`, where `.` is the current working directory
(which should be the main GHG folder for your project).
}
